Add encode/decode options to createPersistentStore and createPersistentMap
Motivation: only primitive values can be stored in localStorage. If I want to persist some complex value like and array of object, I need a way to convert it to string and back.
encode and decode optional functions processes the value before storing it in the storageEngine and after reading the value from storageEngine.